mobile: Remove redundant app-level user checks.
Removed unnecessary user checks. Done with RLS policies and default auth.uid() value in user_id column where needed.

+++ mobile/database/gym/delete-exercise.ts
@@ -1,7 +1,6 @@
 import { supabase } from "@/lib/supabase";
 import { handleError } from "@/utils/handleError";
 
-
 export default async function DeleteExercise(item_id: string) {
   const {
     data: { session },
@@ -15,6 +14,7 @@ export default async function DeleteExercise(item_id: string) {
   const { error } = await supabase
     .from("gym_exercises")
     .delete()
+    .eq("user_id", session.user.id)
     .eq("id", item_id);
 
   if (error) {
